Rearrange
We need to stop defaulting room's privacy to "private", even if things aren't determined.
