+++ src/components/Conversation/index.tsx
@@ -5,7 +5,8 @@ import ConversationHeader from './ConversationHeader'
 import EmptyMessageFeed from './EmptyMessageFeed'
 import MessageFeed from './MessageFeed'
 import MessageInput from './MessageInput'
-import { usePrivacy, useSelectedRoomId } from '$/features/room/hooks'
+import { useSelectedRoomId } from '$/features/room/hooks'
+import usePrivacy from '$/hooks/usePrivacy'
 import useCanGrant from '$/hooks/useCanGrant'
 import useResendEffect from '$/hooks/useResendEffect'
 import useResends from '$/hooks/useResends'
