+++ src/components/modals/EditMembersModal.tsx
@@ -5,7 +5,8 @@ import { Address } from '$/types'
 import { PermissionsAction } from '$/features/permissions'
 import useRoomMembers from '$/hooks/useRoomMembers'
 import useIsDetectingRoomMembers from '$/hooks/useIsDetectingRoomMembers'
-import { usePrivacyOption, useSelectedRoomId } from '$/features/room/hooks'
+import { useSelectedRoomId } from '$/features/room/hooks'
+import usePrivacyOption from '$/hooks/usePrivacyOption'
 import { useWalletAccount, useWalletClient, useWalletProvider } from '$/features/wallet/hooks'
 import useCopy from '$/hooks/useCopy'
 import CopyIcon from '$/icons/CopyIcon'
