+++ src/hooks/usePromoteDelegatedAccount.ts
@@ -1,7 +1,8 @@
 import { useDelegatedAccount } from '$/features/delegation/hooks'
 import { Flag } from '$/features/flag/types'
 import { PermissionsAction } from '$/features/permissions'
-import { usePrivacy, useSelectedRoomId } from '$/features/room/hooks'
+import { useSelectedRoomId } from '$/features/room/hooks'
+import usePrivacy from '$/hooks/usePrivacy'
 import { useWalletAccount, useWalletClient, useWalletProvider } from '$/features/wallet/hooks'
 import { PrivacySetting } from '$/types'
 import { useCallback } from 'react'
